Move PBXReference to files package
